+++ tests/test_caltrack_usage_per_day.py
@@ -798,8 +798,9 @@ def test_get_total_degree_day_too_low_warning_ok():
         model_type="model_type",
         balance_point=65,
         degree_day_type="xdd",
-        degree_days=pd.Series([1, 1, 1]),
-        minimum_total=2,
+        avg_degree_days=pd.Series([1, 1, 1]),
+        period_days=pd.Series([3, 1, 2]),
+        minimum_total=4,
     )
     assert warnings == []
 
@@ -809,8 +810,9 @@ def test_get_total_degree_day_too_low_warning_fail():
         model_type="model_type",
         balance_point=65,
         degree_day_type="xdd",
-        degree_days=pd.Series([0.5, 0.5, 0.5]),
-        minimum_total=2,
+        avg_degree_days=pd.Series([0.5, 0.5, 0.5]),
+        period_days=pd.Series([3, 1, 2]),
+        minimum_total=4,
     )
     assert len(warnings) == 1
     warning = warnings[0]
@@ -821,8 +823,8 @@ def test_get_total_degree_day_too_low_warning_fail():
         "Total XDD below accepted minimum. Candidate fit not attempted."
     )
     assert warning.data == {
-        "total_xdd": 1.5,
-        "total_xdd_minimum": 2,
+        "total_xdd": 3.0,
+        "total_xdd_minimum": 4,
         "xdd_balance_point": 65,
     }
 
@@ -1481,7 +1483,7 @@ def test_fit_caltrack_usage_per_day_model_cdd_hdd_use_billing_presets(
     cdd_hdd_h60_c65, prediction_index, temperature_data
 ):
     model_results = fit_caltrack_usage_per_day_model(
-        cdd_hdd_h60_c65, use_billing_presets=True, weights_col='n_days_kept'
+        cdd_hdd_h60_c65, use_billing_presets=True, weights_col="n_days_kept"
     )
     assert len(model_results.candidates) == 4
     assert model_results.candidates[0].model_type == "intercept_only"
@@ -1498,18 +1500,9 @@ def test_fit_caltrack_usage_per_day_model_cdd_hdd_use_billing_presets(
 def test_fit_caltrack_usage_per_day_model_cdd_hdd_use_billing_presets_no_weights(
     cdd_hdd_h60_c65, prediction_index, temperature_data
 ):
-    model_results = fit_caltrack_usage_per_day_model(
-        cdd_hdd_h60_c65, use_billing_presets=True
-    )
-    assert model_results.status == "ERROR"
-    assert model_results.method_name == "caltrack_usage_per_day"
-    assert model_results.interval == "billing"
-    assert len(model_results.warnings) == 1
-    warning = model_results.warnings[0]
-    assert warning.qualified_name == ("eemeter.caltrack_usage_per_day.missing_weights")
-    assert warning.description == ("Attempting to use billing presets without"
-                                   " providing the weights_col arg.")
-    assert warning.data == {}
+    with pytest.raises(ValueError) as exc_info:
+        fit_caltrack_usage_per_day_model(cdd_hdd_h60_c65, use_billing_presets=True)
+    assert "weights_col" in str(exc_info.value)
 
 
 # When model is intercept-only, num_parameters should = 0 with cvrmse = cvrmse_adj